rid-macro: adding ability to skip a struct field (#26)
- introduced `#[derive(rid::Config)]` which adds rid config attributes
  to the model struct it is applied to
- `#[derive(rid::Config)]` itself does not add any generated code, but
  just serves to introduce the attrs
- those attrs are used to configure each struct field if needed, so far
  `#[rid(skip)]` to omit it from rendering FFI wrapper code
- attr syntax modeled after
  [serde(skip)](https://serde.rs/field-attrs.html#skip)
- as with serde attrs `rid(*)` are extensible for future needs

NOTE: I tried to opaquely add this `derive` to each model struct, but
caused problems when another proc_macro attribute comes after (derive
attrs need to be last). In the future we could explore rewriting the
TokenStream and inject the derive in the right place so the user doesn't
have to add it explicitly.